viewwer

+++ src/boolean_polyline.rs
@@ -22,7 +22,7 @@ struct VOutPt { pt: Pt, next: usize, prev: usize, outrec: usize, horz_used: bool
 struct VOutRec { idx: usize, front_edge: OptIdx, back_edge: OptIdx, pts: OptIdx, owner: OptIdx }
 #[derive(Clone, Default)]
 struct VActive {
-    bot: Pt, top: Pt, curr_x: i64, dx: f32,
+    bot: Pt, top: Pt, curr_x: i64, dx: f64,
     wind_dx: i32, wind_cnt: i32, wind_cnt2: i32,
     outrec: OptIdx, prev_in_ael: OptIdx, next_in_ael: OptIdx,
     prev_in_sel: OptIdx, next_in_sel: OptIdx, jump: OptIdx,
@@ -104,17 +104,17 @@ impl Sc {
 
 // ── Geometry helpers ─────────────────────────────────────────────────────
 
-fn get_dx(p1: Pt, p2: Pt) -> f32 {
-    let dy = (p2.y - p1.y) as f32;
-    if dy != 0.0 { (p2.x - p1.x) as f32 / dy }
-    else if p2.x > p1.x { -f32::MAX } else { f32::MAX }
+fn get_dx(p1: Pt, p2: Pt) -> f64 {
+    let dy = (p2.y - p1.y) as f64;
+    if dy != 0.0 { (p2.x - p1.x) as f64 / dy }
+    else if p2.x > p1.x { -f64::MAX } else { f64::MAX }
 }
 
 fn top_x(sc: &Sc, e: usize, y: i64) -> i64 {
     let a = &sc.act[e];
     if y == a.top.y || a.top.x == a.bot.x { a.top.x }
     else if y == a.bot.y { a.bot.x }
-    else { a.bot.x + (a.dx * (y - a.bot.y) as f32).round() as i64 }
+    else { a.bot.x + (a.dx * (y - a.bot.y) as f64).round() as i64 }
 }
 
 fn is_horz(sc: &Sc, e: usize) -> bool { sc.act[e].top.y == sc.act[e].bot.y }
@@ -136,56 +136,56 @@ fn prev_prev_vertex(sc: &Sc, e: usize) -> OptIdx {
     if sc.act[e].wind_dx > 0 { sc.vtx[sc.vtx[vt].prev?].prev } else { sc.vtx[sc.vtx[vt].next?].next }
 }
 
-fn cross3(p1: Pt, p2: Pt, p3: Pt) -> f32 {
-    (p2.x-p1.x) as f32 * (p3.y-p2.y) as f32 - (p2.y-p1.y) as f32 * (p3.x-p2.x) as f32
+fn cross3(p1: Pt, p2: Pt, p3: Pt) -> f64 {
+    (p2.x-p1.x) as f64 * (p3.y-p2.y) as f64 - (p2.y-p1.y) as f64 * (p3.x-p2.x) as f64
 }
-fn dot3(p1: Pt, p2: Pt, p3: Pt) -> f32 {
-    (p2.x-p1.x) as f32 * (p3.x-p2.x) as f32 + (p2.y-p1.y) as f32 * (p3.y-p2.y) as f32
+fn dot3(p1: Pt, p2: Pt, p3: Pt) -> f64 {
+    (p2.x-p1.x) as f64 * (p3.x-p2.x) as f64 + (p2.y-p1.y) as f64 * (p3.y-p2.y) as f64
 }
 fn is_collinear(p1: Pt, s: Pt, p2: Pt) -> bool {
     (s.x-p1.x) as i128 * (p2.y-s.y) as i128 == (s.y-p1.y) as i128 * (p2.x-s.x) as i128
 }
-fn perpendic_dist_sq(pt: Pt, l1: Pt, l2: Pt) -> f32 {
-    let (a, b, c, d) = ((pt.x-l1.x) as f32, (pt.y-l1.y) as f32, (l2.x-l1.x) as f32, (l2.y-l1.y) as f32);
+fn perpendic_dist_sq(pt: Pt, l1: Pt, l2: Pt) -> f64 {
+    let (a, b, c, d) = ((pt.x-l1.x) as f64, (pt.y-l1.y) as f64, (l2.x-l1.x) as f64, (l2.y-l1.y) as f64);
     if c == 0.0 && d == 0.0 { return 0.0; }
     let e = a*d - c*b; (e*e) / (c*c + d*d)
 }
 
 fn get_seg_isect(a: Pt, b: Pt, c: Pt, d: Pt) -> Option<Pt> {
-    let (dx1, dy1) = ((b.x-a.x) as f32, (b.y-a.y) as f32);
-    let (dx2, dy2) = ((d.x-c.x) as f32, (d.y-c.y) as f32);
+    let (dx1, dy1) = ((b.x-a.x) as f64, (b.y-a.y) as f64);
+    let (dx2, dy2) = ((d.x-c.x) as f64, (d.y-c.y) as f64);
     let det = dy1*dx2 - dy2*dx1;
     if det == 0.0 { return None; }
-    let t = ((a.x-c.x) as f32 * dy2 - (a.y-c.y) as f32 * dx2) / det;
+    let t = ((a.x-c.x) as f64 * dy2 - (a.y-c.y) as f64 * dx2) / det;
     if t <= 0.0 { Some(a) } else if t >= 1.0 { Some(b) }
     else { Some(Pt { x: a.x + (t*dx1).round() as i64, y: a.y + (t*dy1).round() as i64 }) }
 }
 
 fn closest_on_seg(pt: Pt, s1: Pt, s2: Pt) -> Pt {
     if s1 == s2 { return s1; }
-    let (dx, dy) = ((s2.x-s1.x) as f32, (s2.y-s1.y) as f32);
-    let mut q = ((pt.x-s1.x) as f32 * dx + (pt.y-s1.y) as f32 * dy) / (dx*dx + dy*dy);
+    let (dx, dy) = ((s2.x-s1.x) as f64, (s2.y-s1.y) as f64);
+    let mut q = ((pt.x-s1.x) as f64 * dx + (pt.y-s1.y) as f64 * dy) / (dx*dx + dy*dy);
     if q < 0.0 { q = 0.0; } else if q > 1.0 { q = 1.0; }
     Pt { x: s1.x + (q*dx).round() as i64, y: s1.y + (q*dy).round() as i64 }
 }
 
 fn segs_intersect(a: Pt, b: Pt, c: Pt, d: Pt) -> bool {
-    let sign = |v: f32| -> i32 { if v == 0.0 { 0 } else if v > 0.0 { 1 } else { -1 } };
+    let sign = |v: f64| -> i32 { if v == 0.0 { 0 } else if v > 0.0 { 1 } else { -1 } };
     (sign(cross3(a,c,d)) * sign(cross3(b,c,d)) < 0) && (sign(cross3(c,a,b)) * sign(cross3(d,a,b)) < 0)
 }
 
-fn area_outpt(sc: &Sc, start: usize) -> f32 {
+fn area_outpt(sc: &Sc, start: usize) -> f64 {
     let mut r = 0.0; let mut o = start;
     loop {
         let prev = sc.opt[o].prev; let pp = sc.opt[prev].pt; let cp = sc.opt[o].pt;
-        r += (pp.y + cp.y) as f32 * (pp.x - cp.x) as f32;
+        r += (pp.y + cp.y) as f64 * (pp.x - cp.x) as f64;
         o = sc.opt[o].next; if o == start { break; }
     }
     r * 0.5
 }
 
-fn area_tri(p1: Pt, p2: Pt, p3: Pt) -> f32 {
-    (p3.y+p1.y) as f32 * (p3.x-p1.x) as f32 + (p1.y+p2.y) as f32 * (p1.x-p2.x) as f32 + (p2.y+p3.y) as f32 * (p2.x-p3.x) as f32
+fn area_tri(p1: Pt, p2: Pt, p3: Pt) -> f64 {
+    (p3.y+p1.y) as f64 * (p3.x-p1.x) as f64 + (p1.y+p2.y) as f64 * (p1.x-p2.x) as f64 + (p2.y+p3.y) as f64 * (p2.x-p3.x) as f64
 }
 
 fn pts_close(a: Pt, b: Pt) -> bool { (a.x-b.x).abs() < 2 && (a.y-b.y).abs() < 2 }
@@ -224,7 +224,7 @@ fn pip_vertex(sc: &Sc, pt: Pt, head: usize) -> bool {
 
 // ── Vertex building ──────────────────────────────────────────────────────
 
-fn add_path_from_coords(sc: &mut Sc, ca: &[f32], n: usize, pty: i8, bool_scale: f32) -> (OptIdx, i64, i64, i64, i64) {
+fn add_path_from_coords(sc: &mut Sc, ca: &[f64], n: usize, pty: i8, bool_scale: f64) -> (OptIdx, i64, i64, i64, i64) {
     if n < 3 { return (None, 0, 0, 0, 0); }
     let base = sc.vtx.len();
     sc.vtx.resize(base + n, VVertex::default());
@@ -872,8 +872,8 @@ fn insert_local_minima_into_ael(sc: &mut Sc, bot_y: i64, ct: i32) {
         sc.act[rb].local_min = Some(lm_idx); sc.act[rb].dx = get_dx(sc.act[rb].bot, sc.act[rb].top);
         // swap if needed
         let mut lb = lb; let mut rb = rb;
-        if is_horz(sc, lb) { if sc.act[lb].dx == -f32::MAX { std::mem::swap(&mut lb, &mut rb); } }
-        else if is_horz(sc, rb) { if sc.act[rb].dx == f32::MAX { std::mem::swap(&mut lb, &mut rb); } }
+        if is_horz(sc, lb) { if sc.act[lb].dx == -f64::MAX { std::mem::swap(&mut lb, &mut rb); } }
+        else if is_horz(sc, rb) { if sc.act[rb].dx == f64::MAX { std::mem::swap(&mut lb, &mut rb); } }
         else if sc.act[lb].dx < sc.act[rb].dx { std::mem::swap(&mut lb, &mut rb); }
         sc.act[lb].is_left_bound = true;
         insert_left_edge(sc, lb);
@@ -1048,7 +1048,7 @@ fn execute_internal(sc: &mut Sc, ct: i32) -> bool {
 
 /// Returns +1 if convex CCW, -1 if convex CW, 0 if non-convex or degenerate.
 /// Reads x,y from stride-3 coords; ignores z.
-fn convex_orientation_2d(coords: &[f32], n: usize) -> i32 {
+fn convex_orientation_2d(coords: &[f64], n: usize) -> i32 {
     if n < 3 { return 0; }
     let mut sign: i32 = 0;
     for i in 0..n {
@@ -1079,10 +1079,10 @@ fn convex_orientation_2d(coords: &[f32], n: usize) -> i32 {
 ///
 /// Callers must pass thread-local scratch buffers to avoid per-call allocation.
 fn sutherland_hodgman_ccw_intersect(
-    buf_in: &mut Vec<(f32, f32)>,
-    buf_out: &mut Vec<(f32, f32)>,
-    ca: &[f32], na: usize,
-    cb: &[f32], nb: usize,
+    buf_in: &mut Vec<(f64, f64)>,
+    buf_out: &mut Vec<(f64, f64)>,
+    ca: &[f64], na: usize,
+    cb: &[f64], nb: usize,
 ) -> Vec<Polyline> {
     buf_in.clear();
     buf_out.clear();
@@ -1140,7 +1140,7 @@ fn sutherland_hodgman_ccw_intersect(
         // `buf_out` the new subject polygon for the next iteration while
         // `buf_in` becomes the scratch for the next output. Both stay owned
         // by the caller's thread-local cell; capacities persist.
-        std::mem::swap::<Vec<(f32, f32)>>(buf_in, buf_out);
+        std::mem::swap::<Vec<(f64, f64)>>(buf_in, buf_out);
         bx_prev = bx_curr;
         by_prev = by_curr;
     }
@@ -1173,7 +1173,7 @@ thread_local! {
     /// Reusable ping-pong buffers for the Sutherland-Hodgman convex fast path.
     /// Held as a tuple so a single `borrow_mut()` hands out both halves without
     /// running afoul of RefCell aliasing.
-    static SH_BUFFERS: std::cell::RefCell<(Vec<(f32, f32)>, Vec<(f32, f32)>)>
+    static SH_BUFFERS: std::cell::RefCell<(Vec<(f64, f64)>, Vec<(f64, f64)>)>
         = std::cell::RefCell::new((Vec::new(), Vec::new()));
 }
 
@@ -1187,12 +1187,12 @@ pub fn boolean_op(a: &Polyline, b: &Polyline, clip_type: i32) -> Vec<Polyline> {
 
     // Compute safe scale from actual coordinate range to prevent int64 overflow
     // in cross products: (max_coord * scale)^2 must fit in int64
-    let mut max_coord: f32 = 0.0;
+    let mut max_coord: f64 = 0.0;
     for i in 0..na { max_coord = max_coord.max(ca[i*3].abs()).max(ca[i*3+1].abs()); }
     for i in 0..nb { max_coord = max_coord.max(cb[i*3].abs()).max(cb[i*3+1].abs()); }
     if max_coord < 1e-12 { max_coord = 1.0; }
-    let bool_scale: f32 = ((i64::MAX as f32).sqrt() / max_coord * 0.99).floor();
-    let bool_inv_scale: f32 = 1.0 / bool_scale;
+    let bool_scale: f64 = ((i64::MAX as f64).sqrt() / max_coord * 0.99).floor();
+    let bool_inv_scale: f64 = 1.0 / bool_scale;
 
     SCRATCH.with(|cell| {
         let mut sc = cell.borrow_mut();
@@ -1213,14 +1213,14 @@ pub fn boolean_op(a: &Polyline, b: &Polyline, clip_type: i32) -> Vec<Polyline> {
 fn boolean_op_inner(
     sc: &mut Sc,
     a: &Polyline, b: &Polyline,
-    ca: &[f32], cb: &[f32],
+    ca: &[f64], cb: &[f64],
     na: usize, nb: usize,
     clip_type: i32,
-    bool_scale: f32, bool_inv_scale: f32,
+    bool_scale: f64, bool_inv_scale: f64,
 ) -> Vec<Polyline> {
     // Small polygon fast path
     if (na as i64) * (nb as i64) <= 400 {
-        let cvt = |c: &[f32], i: usize| -> Pt { Pt { x: (c[i*3]*bool_scale).round() as i64, y: (c[i*3+1]*bool_scale).round() as i64 } };
+        let cvt = |c: &[f64], i: usize| -> Pt { Pt { x: (c[i*3]*bool_scale).round() as i64, y: (c[i*3+1]*bool_scale).round() as i64 } };
         let va: Vec<Pt> = (0..na).map(|i| cvt(ca, i)).collect();
         let vb: Vec<Pt> = (0..nb).map(|i| cvt(cb, i)).collect();
         let (mut aminx, mut amaxx, mut aminy, mut amaxy) = (va[0].x, va[0].x, va[0].y, va[0].y);
@@ -1333,12 +1333,12 @@ fn boolean_op_inner(
         if very_small_tri(sc, op) { continue; }
         let mut coords = Vec::new();
         let mut o = sc.opt[op].next; let mut last = sc.opt[o].pt;
-        coords.push(last.x as f32 * bool_inv_scale); coords.push(last.y as f32 * bool_inv_scale); coords.push(0.0);
+        coords.push(last.x as f64 * bool_inv_scale); coords.push(last.y as f64 * bool_inv_scale); coords.push(0.0);
         o = sc.opt[o].next;
         let start = sc.opt[op].next;
         while o != start {
             if sc.opt[o].pt != last { last = sc.opt[o].pt;
-                coords.push(last.x as f32 * bool_inv_scale); coords.push(last.y as f32 * bool_inv_scale); coords.push(0.0); }
+                coords.push(last.x as f64 * bool_inv_scale); coords.push(last.y as f64 * bool_inv_scale); coords.push(0.0); }
             o = sc.opt[o].next;
         }
         if coords.len() >= 9 { out.push(Polyline::from_coords(coords)); }
